# docs/api.yaml
openapi: "3.0.2"
info:
  title: Core
  version: 1.0.0
  description: Endpoints for interacting with Core
  contact:
    name: Sebastian Holmgaard
    url: https://github.com/MatchAlly/core
    email: sebholmgaard@gmail.com
servers:
  - url: https://api-matchally.koyeb.app/
  - url: "{scheme}://{host}{base}"
    description: Custom
    variables:
      scheme:
        default: "http"
        enum:
          - "http"
          - "https"
      host:
        default: "localhost:8080"
      base:
        default: ""
...
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Bearer token from MatchAlly"
paths:
...
  auth/refresh:
    post:
      operationId: Refresh
      tags:
        - Authorization endpoints
      description: |
        Endpoint for users refresh their access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
      responses:
        "200":
          description: "Refresh successful"
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
                  refreshToken:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        "400":
          description: "Bad Request"
        "401":
          description: "Unauthorized"
        "500":
          description: "Internal Server Error"

  # User endpoints
...